+++ src/register_types.cpp
@@ -2,6 +2,7 @@
 #include "sentry/disabled/disabled_event.h"
 #include "sentry/dotnet/dotnet_before_send_processor.h"
 #include "sentry/dotnet/dotnet_scope_observer.h"
+#include "sentry/engine_lifecycle/sentry_scene_tree_watcher.h"
 #include "sentry/logging/sentry_godot_logger.h"
 #include "sentry/processing/screenshot_processor.h"
 #include "sentry/processing/sentry_event_processor.h"
@@ -101,6 +102,7 @@ void register_runtime_classes() {
 	GDREGISTER_INTERNAL_CLASS(SentryScopeObserver);
 	GDREGISTER_INTERNAL_CLASS(sentry::dotnet::DotnetScopeObserver);
 	GDREGISTER_INTERNAL_CLASS(sentry::dotnet::DotnetBeforeSendProcessor);
+	GDREGISTER_INTERNAL_CLASS(sentry::engine_lifecycle::SentrySceneTreeWatcher);
 
 #ifdef SDK_NATIVE
 	GDREGISTER_INTERNAL_CLASS(native::NativeEvent);